Use system-font-paths for font export

+++ app_src/lib/fonts/helpers.js
@@ -1,18 +1,41 @@
 const fs = require('fs');
 const path = require('path');
-const fontManager = require('fontmanager-redux');
+const systemFontPaths = require('system-font-paths');
+const fontkit = require('fontkit');
 
-function findFontFile(psName) {
-  const f = fontManager.findFontSync({ postscriptName: psName });
-  return f && f.path ? f.path : null;
+let fontMapPromise;
+
+async function getFontMap() {
+  if (!fontMapPromise) {
+    fontMapPromise = systemFontPaths().then((paths) => {
+      const map = {};
+      paths.forEach((p) => {
+        try {
+          const font = fontkit.openSync(p);
+          if (font && font.postscriptName) {
+            map[font.postscriptName] = p;
+          }
+        } catch (e) {
+          /* ignore unreadable fonts */
+        }
+      });
+      return map;
+    });
+  }
+  return fontMapPromise;
+}
+async function findFontFile(psName) {
+  const map = await getFontMap();
+  return map[psName] || null;
 }
 
-function copyFonts(psNames, destDir) {
+async function copyFonts(psNames, destDir) {
+  const map = await getFontMap();
   if (!fs.existsSync(destDir)) fs.mkdirSync(destDir, { recursive: true });
   const copied = [];
-  psNames.forEach(ps => {
-    const src = findFontFile(ps);
-    if (src) {
+  psNames.forEach((ps) => {
+    const src = map[ps];
+    if (src && fs.existsSync(src)) {
       const target = path.join(destDir, path.basename(src));
       fs.copyFileSync(src, target);
       copied.push(target);
